+++ packages/orq-rc/RELEASES.md
@@ -9549,3 +9549,13 @@ Based on:
 - [typescript v4.10.0-rc.53] packages/orq-rc
 ### Releases
 - [NPM v4.10.0-rc.53] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.53 - packages/orq-rc
+
+## 2026-05-29 07:46:25
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.768.2 (2.889.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.10.0-rc.54] packages/orq-rc
+### Releases
+- [NPM v4.10.0-rc.54] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.54 - packages/orq-rc